fix(web): 迁移首页 locale 根布局
让英文根路径与中文路径分别输出正确的 html lang，迁移旧 /en/ 首页，并移除无公开价值的 kit 路由。
