+++ apps/web/public/_redirects
@@ -1,1 +1,1 @@
-/        /en/    200
+/en/     /       302
