+++ apps/web/src/app/(legal)/layout.tsx
@@ -1,0 +1,9 @@
+import { DocumentShell } from "../_shared/DocumentShell";
+
+export default function LegalLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <DocumentShell lang="en">{children}</DocumentShell>;
+}
