+++ apps/web/src/app/[locale]/page.tsx
@@ -1,26 +1,10 @@
 import type { Metadata } from "next";
-import { getTranslations, setRequestLocale } from "next-intl/server";
+import { setRequestLocale } from "next-intl/server";
+import zhMessages from "../../../messages/zh.json";
+import { notFound } from "next/navigation";
 
-import { AgentIntelligence } from "@/components/sections/AgentIntelligence";
-import { CoreWedge } from "@/components/sections/CoreWedge";
-import { CTAFooter } from "@/components/sections/CTAFooter";
-import { FAQ } from "@/components/sections/FAQ";
-import { GlobalCoverage } from "@/components/sections/GlobalCoverage";
-import { Hero } from "@/components/sections/Hero";
-import { OverfittingGuard } from "@/components/sections/OverfittingGuard";
-import { ResearchFloor } from "@/components/sections/ResearchFloor";
-import { StrategyEvolution } from "@/components/sections/StrategyEvolution";
-import { TrustBoundary } from "@/components/sections/TrustBoundary";
-import { UnifiedKernel } from "@/components/sections/UnifiedKernel";
-import { TickerStrip } from "@/components/primitives/TickerStrip";
-import { getGithubStats } from "@/lib/github";
-import { REPO_COORDS } from "@/lib/links";
-import { getUpdateTag } from "@/lib/release-meta";
-import {
-  buildHomeMetadata,
-  buildHomeStructuredData,
-  isSupportedLocale,
-} from "@/lib/seo";
+import HomePage from "../_shared/HomePage";
+import { buildHomeMetadata } from "@/lib/seo";
 
 export async function generateMetadata({
   params,
@@ -28,72 +12,26 @@ export async function generateMetadata({
   params: Promise<{ locale: string }>;
 }): Promise<Metadata> {
   const { locale } = await params;
-  if (!isSupportedLocale(locale)) {
+  if (locale !== "zh") {
     return {};
   }
 
-  const t = await getTranslations({ locale, namespace: "meta" });
-  return buildHomeMetadata(locale, t("title"), t("description"));
+  return buildHomeMetadata(
+    "zh",
+    zhMessages.meta.title,
+    zhMessages.meta.description,
+  );
 }
 
-export default async function HomePage({
+export default async function ChineseHomePage({
   params,
 }: {
   params: Promise<{ locale: string }>;
 }) {
   const { locale } = await params;
-  if (!isSupportedLocale(locale)) {
-    return null;
+  if (locale !== "zh") {
+    notFound();
   }
 
-  setRequestLocale(locale);
-  const githubStats = await getGithubStats({
-    owner: REPO_COORDS.owner,
-    repo: REPO_COORDS.name,
-  });
-  const updateTag = await getUpdateTag();
-  const tickerItems = [
-    "INALPHA",
-    "OPEN-SOURCE QUANT AGENT FRAMEWORK",
-    updateTag,
-    "AN ORACLE THAT KEEPS A LEDGER",
-    "FACTOR TIMING · RANK IC",
-    "INVESTING LEGENDS PANEL",
-    "STRATEGY EVOLUTION",
-    "ANTI-OVERFITTING · CPCV · DEFLATED SHARPE",
-    "79-FACTOR LIBRARY · DECAY WATCH",
-    "PLAN · APPROVE · EXECUTE",
-    "POSITION GUARD · −20% HARD STOP",
-    "AUTONOMOUS PAPER RUNNER",
-    "AGENTS · FIRST-CLASS",
-    "AGPL-3.0",
-    "12 MARKETS",
-    "INARI OMIKUJI",
-  ];
-  const structuredData = buildHomeStructuredData(locale);
-
-  return (
-    <>
-      <script
-        type="application/ld+json"
-        dangerouslySetInnerHTML={{ __html: JSON.stringify(structuredData) }}
-      />
-      <div className="relative min-h-screen grain bg-bg text-fg">
-        <TickerStrip items={tickerItems} />
-        <Hero />
-        <main className="mx-auto max-w-[96rem] space-y-28 px-6 py-24 md:space-y-40 md:px-14 md:py-28">
-          <CoreWedge />
-          <AgentIntelligence />
-          <ResearchFloor />
-          <StrategyEvolution />
-          <OverfittingGuard />
-          <UnifiedKernel />
-          <TrustBoundary />
-          <GlobalCoverage />
-          <FAQ />
-        </main>
-        <CTAFooter stats={githubStats} updateTag={updateTag} />
-      </div>
-    </>
-  );
+  return <HomePage locale="zh" />;
 }
