+++ apps/web/src/app/_shared/DocumentShell.tsx
@@ -1,0 +1,37 @@
+import "../globals.css";
+
+import { GeistMono } from "geist/font/mono";
+import { GeistSans } from "geist/font/sans";
+import { Fraunces } from "next/font/google";
+
+const fraunces = Fraunces({
+  variable: "--font-display",
+  subsets: ["latin"],
+  weight: ["300", "400", "500", "600"],
+  style: ["normal", "italic"],
+  display: "swap",
+});
+
+const THEME_INIT = `(function(){try{var t=localStorage.getItem('inalpha-theme');if(t!=='light'&&t!=='dark'){t=matchMedia('(prefers-color-scheme: light)').matches?'light':'dark';}var r=document.documentElement;r.setAttribute('data-theme',t);r.style.colorScheme=t;}catch(e){}})();`;
+
+export function DocumentShell({
+  children,
+  lang,
+}: Readonly<{
+  children: React.ReactNode;
+  lang: "en" | "zh";
+}>) {
+  return (
+    <html
+      lang={lang}
+      data-theme="dark"
+      suppressHydrationWarning
+      className={`${GeistSans.variable} ${GeistMono.variable} ${fraunces.variable}`}
+    >
+      <body className="min-h-screen bg-bg text-fg antialiased">
+        <script dangerouslySetInnerHTML={{ __html: THEME_INIT }} />
+        {children}
+      </body>
+    </html>
+  );
+}
